게시글 좋아요 수정중

+++ src/main/kotlin/com/loopin/api/loopinbackend/domain/post/entity/Post.kt
@@ -14,5 +14,5 @@ class Post(
     val content: String? = null,
 
     @Column(nullable = false)
-    val likeCount: Int = 0
+    val likeCount: Long = 0
 ) : BaseEntity()
